Refactored ridge solvers into separate file, so Cholesky can be used for linear regression
